requirements.txt
flask==3.0.3;
gunicorn==21.2.0;
requests==2.30.0;
lxml==4.9.3;
python-dotenv==1.0.0;
pillow==11.1.0;
zipstream==1.1.4;